Fix insert error
If the result contains an error when finished, the error was never
saved to the database. This was because the UPDATE does not return
a new generated ID.

Since we already have the ID, we can remove the getId call and
pass the ID directly.

Removed the incorrect UT and added IT for finished.